Improve testsuite guess for db_XXX/submitRun
Testsuites have to be specified explicitely now as they are not part of
the json format anymore. You usually do this by using a full URL like
db_XXX/v4/YYYY/submitRun to submit to test suite YYYY; Using the old URL
db_XXXX/submitRun would fallback to a hardcoded default of 'nts'.  Turns
out there is a lot more non-nts jobs out there, so improving the
guessing logic for submissions in the old format is worth it.