Make LNT client print server's error messages
The LNT client just prints error codes. In the case of server 500 errors
this is annoying when there are more than one kind of issues going on,
or if you are trying to debug why your data is not being accepted by the
server.  This adds a proper 404 and 500 error page to LNT, which returns
HTML or JSON. Update the submitted to request a JSON reply and if one is
returned, print that error message.